Get the open source Bullet library more in sync with Playstation SPU version

+++ Extras/GIMPACTBullet/btConcaveConcaveCollisionAlgorithm.cpp
@@ -81,7 +81,7 @@ btConcaveConcaveCollisionAlgorithm::~btConcaveConcaveCollisionAlgorithm()
 void btConcaveConcaveCollisionAlgorithm::clearCache()
 {
     btPersistentManifold* mainfold;
-    for (size_t i=0;i<this->m_mainfoldsPtr.size() ; i++)
+    for (int i=0;i<this->m_mainfoldsPtr.size() ; i++)
     {
     	mainfold = m_mainfoldsPtr[i];
     	m_dispatcher->releaseManifold(mainfold);
